refactor and cleanup 6
