+++ .env.example
@@ -7,6 +7,9 @@ SP1_PROVER=mock
 # Private key with the permission to use the network prover (optional if you use SP1_PROVER=mock)
 SP1_PRIVATE_KEY="PRIVATE_KEY"
 
+# Which type of ethereum testnet to run locally (pow|pos) 
+ETH_TESTNET_TYPE="pow"
+
 # Optional field to specify the revision (in the form of a commit hash) of the SP1 operator
 # If not specified the default revision from justfile will be used when building the operator with `just install-operator`
 # Format: SP1_OPERATOR_REV={a commit hash of `https://github.com/cosmos/sp1-ics07-tendermint`}
