+++ e2e/interchaintestv8/e2esuite/suite.go
@@ -2,16 +2,21 @@ package e2esuite
 
 import (
 	"context"
+	"os"
 
 	dockerclient "github.com/docker/docker/client"
 	"github.com/stretchr/testify/suite"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
 
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+
 	sdkmath "cosmossdk.io/math"
 
 	interchaintest "github.com/strangelove-ventures/interchaintest/v8"
 	"github.com/strangelove-ventures/interchaintest/v8/chain/cosmos"
+	icethereum "github.com/strangelove-ventures/interchaintest/v8/chain/ethereum"
 	"github.com/strangelove-ventures/interchaintest/v8/ibc"
 	"github.com/strangelove-ventures/interchaintest/v8/testreporter"
 
@@ -20,17 +25,27 @@ import (
 	"github.com/srdtrk/solidity-ibc-eureka/e2e/v8/testvalues"
 )
 
+const (
+	TestnetTypePoW = "pow"
+	TestnetTypePoS = "pos"
+)
+
 // TestSuite is a suite of tests that require two chains and a relayer
 type TestSuite struct {
 	suite.Suite
 
-	ChainA       ethereum.Ethereum
-	ChainB       *cosmos.CosmosChain
-	UserB        ibc.Wallet
-	dockerClient *dockerclient.Client
-	network      string
-	logger       *zap.Logger
-	ExecRep      *testreporter.RelayerExecReporter
+	ChainA         ethereum.Ethereum
+	ethTestnetType string
+	ChainB         *cosmos.CosmosChain
+	UserB          ibc.Wallet
+	dockerClient   *dockerclient.Client
+	network        string
+	logger         *zap.Logger
+	ExecRep        *testreporter.RelayerExecReporter
+
+	EthereumLightClientID         string
+	TendermintLightClientID       string
+	LastEtheruemLightClientUpdate uint64
 
 	// proposalIDs keeps track of the active proposal ID for cosmos chains
 	proposalIDs map[string]uint64
@@ -40,36 +55,48 @@ type TestSuite struct {
 func (s *TestSuite) SetupSuite(ctx context.Context) {
 	t := s.T()
 
-	chainSpecs := chainconfig.DefaultChainSpecs
+	icChainSpecs := chainconfig.DefaultChainSpecs
 
-	t.Cleanup(func() {
-		ctx := context.Background()
-		if t.Failed() {
-			_ = s.ChainA.DumpLogs(ctx)
-		}
-		s.ChainA.Destroy(ctx)
-	})
-
-	if len(chainSpecs) != 1 {
-		t.Fatal("TestSuite requires exactly 1 chain spec")
+	s.ethTestnetType = os.Getenv(testvalues.EnvKeyEthTestnetType)
+	switch s.ethTestnetType {
+	case TestnetTypePoW:
+		icChainSpecs = append(icChainSpecs, &interchaintest.ChainSpec{ChainConfig: icethereum.DefaultEthereumAnvilChainConfig("ethereum")})
+	case TestnetTypePoS:
+		kurtosisChain, err := chainconfig.SpinUpKurtosisPoS(ctx) // TODO: Run this in a goroutine and wait for it to be ready
+		s.Require().NoError(err)
+		s.ChainA, err = ethereum.NewEthereum(ctx, kurtosisChain.RPC, &kurtosisChain.BeaconApiClient, kurtosisChain.Faucet)
+		s.Require().NoError(err)
+		t.Cleanup(func() {
+			ctx := context.Background()
+			if t.Failed() {
+				_ = kurtosisChain.DumpLogs(ctx)
+			}
+			kurtosisChain.Destroy(ctx)
+		})
+	default:
+		t.Fatalf("Unknown Ethereum testnet type: %s", s.ethTestnetType)
 	}
 
 	s.logger = zaptest.NewLogger(t)
 	s.dockerClient, s.network = interchaintest.DockerSetup(t)
 
-	cf := interchaintest.NewBuiltinChainFactory(s.logger, chainSpecs)
+	cf := interchaintest.NewBuiltinChainFactory(s.logger, icChainSpecs)
 
 	chains, err := cf.Chains(t.Name())
 	s.Require().NoError(err)
-	s.ChainA, err = ethereum.SpinUpEthereum(ctx)
-	s.Require().NoError(err)
+
 	s.ChainB = chains[0].(*cosmos.CosmosChain)
-
-	s.ExecRep = testreporter.NewNopReporter().RelayerExecReporter(t)
-
 	ic := interchaintest.NewInterchain().
 		AddChain(s.ChainB)
 
+	if s.ethTestnetType == TestnetTypePoW {
+		anvil := chains[1].(*icethereum.EthereumChain)
+		ic = ic.AddChain(anvil)
+	}
+
+	s.ExecRep = testreporter.NewNopReporter().RelayerExecReporter(t)
+
+	// TODO: Run this in a goroutine and wait for it to be ready
 	s.Require().NoError(ic.Build(ctx, s.ExecRep, interchaintest.InterchainBuildOptions{
 		TestName:         t.Name(),
 		Client:           s.dockerClient,
@@ -79,6 +106,15 @@ func (s *TestSuite) SetupSuite(ctx context.Context) {
 
 	// map all query request types to their gRPC method paths for cosmos chains
 	s.Require().NoError(PopulateQueryReqToPath(ctx, s.ChainB))
+
+	if s.ethTestnetType == TestnetTypePoW {
+		anvil := chains[1].(*icethereum.EthereumChain)
+		faucet, err := crypto.ToECDSA(ethcommon.FromHex("0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80"))
+		s.Require().NoError(err)
+
+		s.ChainA, err = ethereum.NewEthereum(ctx, anvil.GetHostRPCAddress(), nil, faucet)
+		s.Require().NoError(err)
+	}
 
 	// Fund user accounts
 	cosmosUserFunds := sdkmath.NewInt(testvalues.InitialBalance)
