+++ e2e/interchaintestv8/ethereum/ethereum.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"io"
 	"math/big"
@@ -13,13 +12,6 @@ import (
 	"os/exec"
 	"strconv"
 	"strings"
-	"time"
-
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/enclaves"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/services"
-	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/starlark_run_config"
-	"github.com/kurtosis-tech/kurtosis/api/golang/engine/kurtosis_engine_rpc_api_bindings"
-	"github.com/kurtosis-tech/kurtosis/api/golang/engine/lib/kurtosis_context"
 
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/crypto"
@@ -27,79 +19,19 @@ import (
 
 	"cosmossdk.io/math"
 
-	"github.com/strangelove-ventures/interchaintest/v8/testutil"
-
 	"github.com/srdtrk/solidity-ibc-eureka/e2e/v8/testvalues"
 )
 
 type Ethereum struct {
-	kurtosisCtx *kurtosis_context.KurtosisContext
-	enclaveCtx  *enclaves.EnclaveContext
-
-	Started         bool
 	ChainID         *big.Int
 	RPC             string
 	EthAPI          EthAPI
-	BeaconAPIClient BeaconAPIClient
+	BeaconAPIClient *BeaconAPIClient
 
 	Faucet *ecdsa.PrivateKey
 }
 
-type KurtosisNetworkParams struct {
-	Participants        []Participant       `json:"participants"`
-	NetworkParams       NetworkConfigParams `json:"network_params"`
-	WaitForFinalization bool                `json:"wait_for_finalization"`
-}
-
-type Participant struct {
-	CLType string `json:"cl_type"`
-	// CLImage       string   `json:"cl_image"`
-	// CLExtraParams []string `json:"cl_extra_params"`
-	ELImage    string `json:"el_image"`
-	ELLogLevel string `json:"el_log_level"`
-}
-
-type NetworkConfigParams struct {
-	Preset string `json:"preset"`
-	// SecondsPerSlot uint64 `json:"seconds_per_slot"`
-}
-
-func ConnectToRunningEthereum(ctx context.Context) (Ethereum, error) {
-	// get faucet private key from string
-	faucet, err := crypto.ToECDSA(ethcommon.FromHex(testvalues.FaucetPrivateKey))
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	kurtosisCtx, err := kurtosis_context.NewKurtosisContextFromLocalEngine()
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	enclaveName := "ethereum-pos-testnet"
-	enclaves, err := kurtosisCtx.GetEnclaves(ctx)
-	if err != nil {
-		return Ethereum{}, err
-	}
-	var enclaveInfo *kurtosis_engine_rpc_api_bindings.EnclaveInfo
-	if enclaveInfos, found := enclaves.GetEnclavesByName()[enclaveName]; found {
-		if len(enclaveInfos) != 1 {
-			return Ethereum{}, fmt.Errorf("expected 1 enclave, found %d", len(enclaveInfos))
-		}
-		enclaveInfo = enclaveInfos[0]
-	}
-
-	enclaveCtx, err := kurtosisCtx.GetEnclaveContext(ctx, enclaveInfo.EnclaveUuid)
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	gethCtx, err := enclaveCtx.GetServiceContext("el-1-geth-lodestar")
-	if err != nil {
-		return Ethereum{}, err
-	}
-	rpcPortSpec := gethCtx.GetPublicPorts()["rpc"]
-	rpc := fmt.Sprintf("http://localhost:%d", rpcPortSpec.GetNumber())
+func NewEthereum(ctx context.Context, rpc string, beaconAPIClient *BeaconAPIClient, faucet *ecdsa.PrivateKey) (Ethereum, error) {
 	ethClient, err := ethclient.Dial(rpc)
 	if err != nil {
 		return Ethereum{}, err
@@ -113,142 +45,7 @@ func ConnectToRunningEthereum(ctx context.Context) (Ethereum, error) {
 		return Ethereum{}, err
 	}
 
-	lighthouseCtx, err := enclaveCtx.GetServiceContext("cl-1-lodestar-geth")
-	if err != nil {
-		return Ethereum{}, err
-	}
-	beaconPortSpec := lighthouseCtx.GetPublicPorts()["http"]
-	beaconRPC := fmt.Sprintf("http://localhost:%d", beaconPortSpec.GetNumber())
-
 	return Ethereum{
-		kurtosisCtx:     kurtosisCtx,
-		enclaveCtx:      enclaveCtx,
-		Started:         true,
-		ChainID:         chainID,
-		RPC:             rpc,
-		EthAPI:          ethAPI,
-		BeaconAPIClient: NewBeaconAPIClient(beaconRPC),
-		Faucet:          faucet,
-	}, nil
-}
-
-func SpinUpEthereum(ctx context.Context) (Ethereum, error) {
-	// get faucet private key from string
-	faucet, err := crypto.ToECDSA(ethcommon.FromHex(testvalues.FaucetPrivateKey))
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	kurtosisCtx, err := kurtosis_context.NewKurtosisContextFromLocalEngine()
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	enclaveName := "ethereum-pos-testnet"
-	enclaves, err := kurtosisCtx.GetEnclaves(ctx)
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	if enclaveInfos, found := enclaves.GetEnclavesByName()[enclaveName]; found {
-		for _, enclaveInfo := range enclaveInfos {
-			err = kurtosisCtx.DestroyEnclave(ctx, enclaveInfo.EnclaveUuid)
-			if err != nil {
-				return Ethereum{}, err
-			}
-		}
-	}
-	enclaveCtx, err := kurtosisCtx.CreateEnclave(ctx, enclaveName)
-	if err != nil {
-		return Ethereum{}, err
-	}
-	networkParams := KurtosisNetworkParams{
-		Participants: []Participant{
-			{
-				CLType: "lodestar",
-				// CLImage:       "sigp/lighthouse:latest-unstable",
-				// CLExtraParams: []string{"--light-client-server"},
-				ELImage:    "ethereum/client-go:v1.14.6",
-				ELLogLevel: "info",
-			},
-		},
-		NetworkParams: NetworkConfigParams{
-			Preset: "minimal",
-		},
-		WaitForFinalization: true,
-	}
-	networkParamsJson, err := json.Marshal(networkParams)
-	if err != nil {
-		return Ethereum{}, err
-	}
-	starlarkResp, err := enclaveCtx.RunStarlarkRemotePackageBlocking(ctx, "github.com/ethpandaops/ethereum-package", &starlark_run_config.StarlarkRunConfig{
-		SerializedParams: string(networkParamsJson),
-	})
-	if err != nil {
-		return Ethereum{}, err
-	}
-	fmt.Println(starlarkResp.RunOutput)
-
-	gethCtx, err := enclaveCtx.GetServiceContext("el-1-geth-lodestar")
-	if err != nil {
-		return Ethereum{}, err
-	}
-	rpcPortSpec := gethCtx.GetPublicPorts()["rpc"]
-	rpc := fmt.Sprintf("http://localhost:%d", rpcPortSpec.GetNumber())
-	ethClient, err := ethclient.Dial(rpc)
-	if err != nil {
-		return Ethereum{}, err
-	}
-	chainID, err := ethClient.ChainID(ctx)
-	if err != nil {
-		return Ethereum{}, err
-	}
-	ethAPI, err := NewEthAPI(rpc)
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	lighthouseCtx, err := enclaveCtx.GetServiceContext("cl-1-lodestar-geth")
-	if err != nil {
-		return Ethereum{}, err
-	}
-	beaconPortSpec := lighthouseCtx.GetPublicPorts()["http"]
-	beaconRPC := fmt.Sprintf("http://localhost:%d", beaconPortSpec.GetNumber())
-
-	beaconAPIClient := NewBeaconAPIClient(beaconRPC)
-	err = testutil.WaitForCondition(10*time.Minute, 5*time.Second, func() (bool, error) {
-		finalizedBlocksResp, err := beaconAPIClient.GetFinalizedBlocks()
-		fmt.Printf("Waiting for chain to finalize, finalizedBlockResp: %+v, err: %s\n", finalizedBlocksResp, err)
-		if err != nil {
-			return false, nil
-		}
-		if !finalizedBlocksResp.Finalized {
-			return false, nil
-		}
-
-		// executionHeight, err := beaconAPIClient.GetExecutionHeight("finalized")
-		// if err != nil {
-		// 	return false, nil
-		// }
-		header, err := beaconAPIClient.GetHeader(finalizedBlocksResp.Data.Message.Slot)
-		if err != nil {
-			return false, nil
-		}
-		bootstrap, err := beaconAPIClient.GetBootstrap(header.Root)
-		if err != nil {
-			return false, nil
-		}
-
-		return bootstrap.Data.Header.Beacon.Slot != 0, nil
-	})
-	if err != nil {
-		return Ethereum{}, err
-	}
-
-	return Ethereum{
-		kurtosisCtx:     kurtosisCtx,
-		enclaveCtx:      enclaveCtx,
-		Started:         true,
 		ChainID:         chainID,
 		RPC:             rpc,
 		EthAPI:          ethAPI,
@@ -257,64 +54,16 @@ func SpinUpEthereum(ctx context.Context) (Ethereum, error) {
 	}, nil
 }
 
-func (e Ethereum) Destroy(ctx context.Context) {
-	if !e.Started {
-		return
-	}
-	// TODO: Turn back on before merging
-	// if err := e.kurtosisCtx.DestroyEnclave(ctx, string(e.enclaveCtx.GetEnclaveUuid())); err != nil {
-	// 	panic(err)
-	// }
-}
-
-func (e Ethereum) DumpLogs(ctx context.Context) error {
-	if !e.Started {
-		return nil
-	}
-	enclaveServices, err := e.enclaveCtx.GetServices()
-	if err != nil {
-		return err
-	}
-
-	userServices := make(map[services.ServiceUUID]bool)
-	serviceIdToName := make(map[services.ServiceUUID]string)
-	for serviceName, servicesUUID := range enclaveServices {
-		userServices[servicesUUID] = true
-		serviceIdToName[servicesUUID] = string(serviceName)
-
-	}
-
-	stream, cancelFunc, err := e.kurtosisCtx.GetServiceLogs(ctx, string(e.enclaveCtx.GetEnclaveUuid()), userServices, false, true, 0, nil)
-	if err != nil {
-		return err
-	}
-
-	// Dump the stream chan into stdout
-	fmt.Println("Dumping kurtosis logs")
-	for {
-		select {
-		case logs, ok := <-stream:
-			if !ok {
-				return nil
-			}
-			for serviceID, serviceLog := range logs.GetServiceLogsByServiceUuids() {
-				if serviceIdToName[serviceID] != "el-1-geth-lodestar" {
-					continue
-				}
-				for _, log := range serviceLog {
-					fmt.Printf("Service %s logs: %s\n", serviceIdToName[serviceID], log)
-				}
-			}
-		case <-ctx.Done():
-			cancelFunc()
-			return nil
-		}
-	}
-}
-
 func (e Ethereum) ForgeScript(deployer *ecdsa.PrivateKey, solidityContract string) ([]byte, error) {
 	cmd := exec.Command("forge", "script", "--rpc-url", e.RPC, "--broadcast", "--non-interactive", "-vvvv", solidityContract)
-	cmd.Env = append(cmd.Env, fmt.Sprintf("PRIVATE_KEY=0x%s", hex.EncodeToString(deployer.D.Bytes())))
+
+	faucetAddress := crypto.PubkeyToAddress(e.Faucet.PublicKey)
+	extraEnv := []string{
+		fmt.Sprintf("%s=%s", testvalues.EnvKeyE2EFacuetAddress, faucetAddress.Hex()),
+		fmt.Sprintf("PRIVATE_KEY=0x%s", hex.EncodeToString(deployer.D.Bytes())),
+	}
+
+	cmd.Env = append(cmd.Env, extraEnv...)
 
 	var stdoutBuf bytes.Buffer
 
@@ -367,7 +116,11 @@ func (e Ethereum) SendEth(key *ecdsa.PrivateKey, toAddress string, amount math.I
 	cmd.Stdout = os.Stdout
 	cmd.Stderr = os.Stderr
 
-	return cmd.Run()
+	if err := cmd.Run(); err != nil {
+		return fmt.Errorf("failed to send eth with %s: %w", strings.Join(cmd.Args, " "), err)
+	}
+
+	return nil
 }
 
 func (e *Ethereum) Height() (int64, error) {
